Istio steps: do not patch privileged SCC
The steps in the README for the istio example included a patch of the
privileged SCC to set "NET_ADMIN" as the list of allowed capabilities.
This is not necessary, as the default privileged SCC already allows
all caps. Moreover, it prevents apps to deploy, as they try to request
CAP_NET_ADMIN (not NET_ADMIN).
Also there was an unneded dir change that makes the rest of the
relative paths fail.